src/backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "crypto-portfolio-backend"
version = "0.1.0"
description = "FastAPI backend for cryptocurrency portfolio tracker"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.120.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "redis[hiredis]>=4.2.0",
    "PyJWT>=2.10.1",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "pyotp>=2.9.0",
    "qrcode[pil]>=8.2",
    "cryptography>=46.0.3",
    "authlib>=1.3.0",
    "aiohttp>=3.9.0",
    "celery>=5.3.0",
    "boto3>=1.26.0",
    "requests>=2.31.0",
    "user-agents>=2.2.0",
    "geoip2>=5.1.0",
    "alembic>=1.13.0",
    "email-validator>=2.1.0",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "httpx>=0.25.0",
    "pytest-cov>=7.0.0",
    "black>=25.9.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
